+++ src/Helper/Charset.php
@@ -3,13 +3,14 @@
 namespace PhpXmlRpc\Helper;
 
 use PhpXmlRpc\PhpXmlRpc;
+use PhpXmlRpc\Traits\LoggerAware;
 
 /**
  * @todo implement an interface
  */
 class Charset
 {
-    protected static $logger;
+    use LoggerAware;
 
     // tables used for transcoding different charsets into us-ascii xml
     protected $xml_iso88591_Entities = array("in" => array(), "out" => array());
@@ -41,23 +42,6 @@ class Charset
         }
 
         return self::$instance;
-    }
-
-    public function getLogger()
-    {
-        if (self::$logger === null) {
-            self::$logger = Logger::instance();
-        }
-        return self::$logger;
-    }
-
-    /**
-     * @param $logger
-     * @return void
-     */
-    public static function setLogger($logger)
-    {
-        self::$logger = $logger;
     }
 
     /**