+++ tests/04LoggerTest.php
@@ -23,7 +23,10 @@ class LoggerTest extends PhpXmlRpc_PolyfillTestCase
         $l = $ch->getLogger();
         Charset::setLogger($this);
 
+        ob_start();
         $ch->encodeEntities('hello world', 'UTF-8', 'NOT-A-CHARSET');
+        $o = ob_get_clean();
+        $this->assertEquals('', $o);
         $this->assertStringContainsString("via mbstring: failed", $this->errorBuffer);
 
         Charset::setLogger($l);
@@ -31,15 +34,18 @@ class LoggerTest extends PhpXmlRpc_PolyfillTestCase
 
     public function testHttpAltLogger()
     {
-        $l = Http::getLogger();
+        $h = new Http();
+        $l = $h->getLogger();
         Http::setLogger($this);
 
-        $h = new Http();
         $s = "HTTP/1.0 200 OK\r\n" .
             "Content-Type: unknown\r\n" .
             "\r\n" .
             "body";
+        ob_start();
         $h->parseResponseHeaders($s, false, 1);
+        $o = ob_get_clean();
+        $this->assertEquals('', $o);
         $this->assertStringContainsString("HEADER: content-type: unknown", $this->debugBuffer);
         Http::setLogger($l);
     }
@@ -50,7 +56,10 @@ class LoggerTest extends PhpXmlRpc_PolyfillTestCase
         $l = $xp->getLogger();
         XMLParser::setLogger($this);
 
+        ob_start();
         $xp->parse('<?xml version="1.0" ?><methodResponse><params><param><value><boolean>x</boolean></value></param></params></methodResponse>');
+        $o = ob_get_clean();
+        $this->assertEquals('', $o);
         $this->assertStringContainsString("invalid data received in BOOLEAN value", $this->errorBuffer);
 
         XMLParser::setLogger($l);